+++ src/components/Notification/styles.css
@@ -54,7 +54,7 @@
   background-color: white;
   border: 1px solid #ddd;
   padding: 20px;
-  width: 300px;
+  width: 350px;
   box-shadow: 0 4px 12px rgba(0, 0, 0, 0.15);
   display: none;
   z-index: 1000;
