+++ src/constants/common.ts
@@ -1,2 +1,4 @@
 export const USERS = 'users';
 export const VALID_EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+export const BID_NOTIFICATIONS = 'BID_NOTIFICATIONS';
+export const LOGGED_IN_USER = 'LOGGED_IN_USER';
