+++ src/types/notification.ts
@@ -1,0 +1,8 @@
+export type Notification = {
+  productId: any;
+  userId: string;
+  userName: string;
+  amount: number;
+  productName: string;
+  timestamp?: number;
+};
